Add Everything SDK wrapper and refactor server file for improved search functionality
- Introduced a new EverythingSDK class in everything_sdk.py to encapsulate Everything SDK functionality, including search capabilities and result handling.
- Refactored server.py to utilize the EverythingSDK for file searching, removing redundant code and improving maintainability.
- Updated search logic to handle search parameters and results more efficiently, enhancing overall performance and clarity.
